Move LcidInfo to v2 folder
