+++ src/views/App/index.tsx
@@ -25,8 +25,8 @@ import GovernanceOverview from 'views/GovernanceOverview';
 import GovernanceReviewTeam from 'views/GovernanceReviewTeam';
 import GovernanceTaskList from 'views/GovernanceTaskList';
 import GovernanceFeedback from 'views/GovernanceTaskList/Feedback';
+import LcidInfo from 'views/GovernanceTaskList/LcidInfo';
 import GovernanceTaskListV1 from 'views/GovernanceTaskListV1';
-import LcidInfo from 'views/GovernanceTaskListV1/LcidInfo';
 import RequestDecision from 'views/GovernanceTaskListV1/RequestDecision';
 import Help from 'views/Help';
 import Home from 'views/Home';
