+++ src/views/GovernanceTaskList/LcidInfo/index.scss
